// client/src/app/models/certificate.ts
export class Certificate{
    id: number;
    event: string;
    date_event: string;
    institution: string;
    ch: number;
    file_url: string;
    status: number;
    description: string;
    student: string;
    administrator: string;
    rule: number;
    created_at: string;
    updated_at: string;
}